[ALPS09080691] dma-buf: follow dma_heap_ops change
[Description]
fix build error according to 30d2da22a338 ("BACKPORT: FROMLIST: dma-buf/heaps: Correct the types of fd_flags and heap_flags") of kernel-6.6

kernel_device_modules-6.6/drivers/dma-buf/heaps/mtk_system_heap.c:1368:14: error: incompatible function pointer types initializing struct dma_buf *(*)(struct dma_heap *, unsigned long, u32, u64) (aka struct dma_buf *(*)(struct dma_heap *, unsigned long, unsigned int, unsigned long long)) with an expression of type struct dma_buf *(struct dma_heap *, unsigned long, unsigned long, unsigned long) [-Wincompatible-function-pointer-types]

[Test]
sanity/stability

MTK-Commit-Id: 7ec6faa419612d064a1016c2ddd726b2b5359d4b

CR-Id: ALPS09080691
Feature: Kernel SI Operation

+++ drivers/dma-buf/heaps/mtk_system_heap.c
@@ -1148,8 +1148,8 @@ static struct page *alloc_largest_available(unsigned long size,
 
 static struct dma_buf *system_heap_do_allocate(struct dma_heap *heap,
 					       unsigned long len,
-					       unsigned long fd_flags,
-					       unsigned long heap_flags,
+					       u32 fd_flags,
+					       u64 heap_flags,
 					       bool uncached,
 					       const struct dma_buf_ops *ops)
 {
@@ -1322,8 +1322,8 @@ free_buffer:
 
 static struct dma_buf *system_heap_allocate(struct dma_heap *heap,
 					    unsigned long len,
-					    unsigned long fd_flags,
-					    unsigned long heap_flags)
+					    u32 fd_flags,
+					    u64 heap_flags)
 {
 	struct mtk_heap_priv_info *heap_priv = dma_heap_get_drvdata(heap);
 
@@ -1334,8 +1334,8 @@ static struct dma_buf *system_heap_allocate(struct dma_heap *heap,
 
 static struct dma_buf *mtk_mm_heap_allocate(struct dma_heap *heap,
 					    unsigned long len,
-					    unsigned long fd_flags,
-					    unsigned long heap_flags)
+					    u32 fd_flags,
+					    u64 heap_flags)
 {
 	struct mtk_heap_priv_info *heap_priv = dma_heap_get_drvdata(heap);
 
@@ -1346,8 +1346,8 @@ static struct dma_buf *mtk_mm_heap_allocate(struct dma_heap *heap,
 
 static struct dma_buf *mtk_slc_heap_allocate(struct dma_heap *heap,
 					    unsigned long len,
-					    unsigned long fd_flags,
-					    unsigned long heap_flags)
+					    u32 fd_flags,
+					    u64 heap_flags)
 {
 	return system_heap_do_allocate(heap, len, fd_flags, heap_flags, false,
 				       &mtk_slc_heap_buf_ops);
